harness/backend/requirements.txt
claude-agent-sdk>=0.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=13.0
watchdog>=5.0
pydantic>=2.8
pyyaml>=6.0
pytest>=8.0
pytest-asyncio>=0.24
httpx>=0.27
